Start refactor of validateProduct
Moves tests for validateProduct method to
ProductControllerValidateProductSpec
