+++ server/src/test/java/umm3601/product/ProductControllerSpec.java
@@ -35,7 +35,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import io.javalin.core.JavalinConfig;
-import io.javalin.core.validation.ValidationException;
 import io.javalin.http.BadRequestResponse;
 import io.javalin.http.Context;
 import io.javalin.http.HandlerType;
@@ -424,269 +423,6 @@ public class ProductControllerSpec {
     assertTrue(addedProduct.containsKey("image"));
   }
 
-  @Test
-  public void addInvalidCategoryProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"\"," // Invalid category
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 10,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addInvalidThresholdProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 0," // Invalid threshold
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addEmptyLifespanProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": null," // empty lifespan
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addNullNameProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": null," // null name
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addEmptyNameProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"\"," // empty name
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addNullDescriptionProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\": null," //null description
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addEmptyProductDescription() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\"," // empty name
-        + "\"description\":\"\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    productController.addNewProduct(ctx);
-    String result = ctx.resultString();
-    String id = javalinJackson.fromJsonString(result, ObjectNode.class).get("id").asText();
-
-    // Our status should be 201, i.e., our new product was successfully
-    // created. This is a named constant in the class HttpURLConnection.
-    assertEquals(HttpURLConnection.HTTP_CREATED, mockRes.getStatus());
-
-    // Successfully adding the product should return the newly generated MongoDB ID
-    // for that product.
-    assertNotEquals("", id);
-    assertEquals(1, db.getCollection("products").countDocuments(eq("_id", new ObjectId(id))));
-
-    // Verify that the product was added to the database with the correct ID
-    Document addedProduct = db.getCollection("products").find(eq("_id", new ObjectId(id))).first();
-
-    assertEquals("", addedProduct.getString("description"));
-  }
-
-  @Test
-  public void addNullBrandProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": null," // empty brand
-        + "\"category\": \"test category\","
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addNullCategoryProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": null ," // empty category
-        + "\"store\": \"test store\","
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": \"test tag\""
-        + "\"lifespan\": 100"
-        + "\"threshold\": 84"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
-
-  @Test
-  public void addNullStoreProduct() throws IOException {
-    String testNewProduct = "{"
-        + "\"product_name\": \"Test Product name\","
-        + "\"description\":\"A test product description\","
-        + "\"brand\": \"test brand\","
-        + "\"category\": \"test category\","
-        + "\"store\": null," // empty store
-        + "\"location\": \"test location\","
-        + "\"notes\": \"tastes like test\","
-        + "\"tags\": [\"test tag\"],"
-        + "\"lifespan\": 100,"
-        + "\"threshold\": 84,"
-        + "\"image\": \"https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon\""
-        + "}";
-    mockReq.setBodyContent(testNewProduct);
-    mockReq.setMethod("POST");
-
-    Context ctx = mockContext("api/products");
-
-    assertThrows(ValidationException.class, () -> {
-      productController.addNewProduct(ctx);
-    });
-  }
 
   @Test
   public void deleteProduct() throws IOException {
